requirements.txt
paramiko>=3.4
pyyaml>=6.0
ruamel.yaml>=0.18
fastapi>=0.115
uvicorn>=0.32
websockets>=13.0
python-multipart>=0.0.9
cryptography>=42.0
